Move context classes into inner classes
