+++ fabric-rendering-v1/src/client/java/net/fabricmc/fabric/api/client/rendering/v1/advancement/AdvancementRenderContext.java
@@ -24,9 +24,10 @@ import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.AdvancementProgress;
 import net.minecraft.advancements.DisplayInfo;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.navigation.ScreenRectangle;
 
 @ApiStatus.NonExtendable
-public interface AbstractAdvancementRenderContext {
+public sealed interface AdvancementRenderContext permits AdvancementRenderContext.Icon, AdvancementRenderContext.Frame, AdvancementRenderContext.Background {
 	/**
 	 * The graphics instance used for rendering.
 	 * @return {@link GuiGraphics} instance
@@ -45,10 +46,18 @@ public interface AbstractAdvancementRenderContext {
 	@Nullable
 	AdvancementProgress progress();
 
+	/**
+	 * The advancement being rendered.
+	 * @return {@link Advancement} instance
+	 */
 	default Advancement advancement() {
 		return holder().value();
 	}
 
+	/**
+	 * The display info of the advancement.
+	 * @return {@link DisplayInfo} instance
+	 */
 	default DisplayInfo display() {
 		return advancement().display().orElseThrow();
 	}
@@ -59,5 +68,65 @@ public interface AbstractAdvancementRenderContext {
 	default boolean isObtained() {
 		AdvancementProgress progress = progress();
 		return progress != null && progress.getPercent() >= 1;
+	}
+
+	@ApiStatus.NonExtendable
+	non-sealed interface Icon extends AdvancementRenderContext {
+		/**
+		 * @return The x coordinate of the icon's top-left corner.
+		 */
+		int x();
+
+		/**
+		 * @return The y coordinate of the icon's top-left corner.
+		 */
+		int y();
+
+		/**
+		 * @return {@code true} if the mouse is hovered over the icon.
+		 */
+		boolean isHovered();
+
+		/**
+		 * @return {@code true} if the icon is rendered as a selected tab.
+		 */
+		boolean isSelected();
+	}
+
+	@ApiStatus.NonExtendable
+	non-sealed interface Frame extends AdvancementRenderContext {
+		/**
+		 * @return The x coordinate of the frame's top-left corner.
+		 */
+		int x();
+
+		/**
+		 * @return The y coordinate of the frame's top-left corner.
+		 */
+		int y();
+
+		/**
+		 * @return {@code true} if the mouse is hovered over the frame.
+		 */
+		boolean isHovered();
+	}
+
+	@ApiStatus.NonExtendable
+	non-sealed interface Background extends AdvancementRenderContext {
+		/**
+		 * @return the {@link ScreenRectangle} that the background is contained within.
+		 * @apiNote use {@link ScreenRectangle#left()} and {@link ScreenRectangle#top()} for the starting coordinates of the background.
+		 */
+		ScreenRectangle bounds();
+
+		/**
+		 * @return the background's x scroll offset.
+		 */
+		double scrollX();
+
+		/**
+		 * @return the background's y scroll offset.
+		 */
+		double scrollY();
 	}
 }
